Update dataset scripts
Previously, the model expected the PTB dataset as input.
This commit generalizes this expectation by now looking for
- train.txt
- test.txt
- valid.txt
in some data folder to be specified (the default is data/).

A new script `data/dictionary_splitter.py` was also added that
can generate subsets of words from the Merriam-Webster Dictionary
(matthewreagan/WebstersEnglishDictionary).